Refactoring
Splitting functions into different files for reuse in new bots
